refactor: rename submodule to abi
